+++ Docs/UltraCanvas/CHANGELOG.md
@@ -1,6 +1,7 @@
 2026-04-16 1.0.17
 - Add on-the-fly submenu regeneration to UltraCanvasMenu
 - Fix Windows high CPU usage when app is idle.
+- Fix main-row digit/punctuation key mappings on Linux and macOS
 
 2026-04-16 1.0.16
 - Full rework of display and rendering the text. Use Pango layout to format text. Allow to use variable height lines